+++ packages/better_player_platform_interface/CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Updated: Updated doc comment for setSpeed to reflect the new 4.0 limit.
+
 ## 1.3.1
 - Updated: package metadata